Crate snowbridge_pallet_inbound_queue

Inbound Queue

Overview

Receives messages emitted by the Gateway contract on Ethereum, whereupon they are verified, translated to XCM, and finally sent to their final destination parachain.

The message relayers are rewarded using native currency from the sovereign account of the destination parachain.

Extrinsics

Governance

Call::set_operating_mode: Set the operating mode of the pallet. Can be used to disable processing of inbound messages.

Message Submission

Call::submit: Submit a message for verification and dispatch the final destination parachain.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for snowbridge_inbound_queue

Constants

LOG_TARGET