Trait snowbridge_pallet_inbound_queue::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Verifier: Verifier;
    type Token: Mutate<Self::AccountId> + Inspect<Self::AccountId>;
    type XcmSender: SendXcm;
    type GatewayAddress: Get<H160>;
    type MessageConverter: ConvertMessage<AccountId = Self::AccountId, Balance = <<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>;
    type ChannelLookup: StaticLookup<Source = ChannelId, Target = Channel>;
    type PricingParameters: Get<PricingParameters<<<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>>;
    type WeightInfo: WeightInfo;
    type Helper: BenchmarkHelper<Self>;
    type WeightToFee: WeightToFee<Balance = <<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>;
    type LengthToFee: WeightToFee<Balance = <<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>;
    type MaxMessageSize: Get<u32>;
    type AssetTransactor: TransactAsset;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type Verifier: Verifier

The verifier for inbound messages from Ethereum

type Token: Mutate<Self::AccountId> + Inspect<Self::AccountId>

Message relayers are rewarded with this asset

type XcmSender: SendXcm

XCM message sender

type GatewayAddress: Get<H160>

type MessageConverter: ConvertMessage<AccountId = Self::AccountId, Balance = <<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>

Convert inbound message to XCM

type ChannelLookup: StaticLookup<Source = ChannelId, Target = Channel>

Lookup a channel descriptor

type PricingParameters: Get<PricingParameters<<<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>>

Lookup pricing parameters

type WeightInfo: WeightInfo

type Helper: BenchmarkHelper<Self>

type WeightToFee: WeightToFee<Balance = <<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>

Convert a weight value into deductible balance type.

type LengthToFee: WeightToFee<Balance = <<Self as Config>::Token as Inspect<<Self as Config>::AccountId>>::Balance>

Convert a length value into deductible balance type

type MaxMessageSize: Get<u32>

The upper limit here only used to estimate delivery cost

type AssetTransactor: TransactAsset

To withdraw and deposit an asset.

Object Safety

This trait is not object safe.